feat(eslint-config-angular): support member ordering
