+++ package.json
@@ -41,8 +41,8 @@
     "@types/eslint": "^7.2.0",
     "@types/jest": "^26.0.0",
     "@types/node": "^14.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.3.0",
-    "@typescript-eslint/parser": "^5.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "babel-jest": "^26.0.1",
     "concurrently": "5.0.1",
     "conventional-changelog-conventionalcommits": "^4.2.3",
