+++ packages/eslint-config-angular/index.js
@@ -1,29 +1,28 @@
 module.exports = {
-    root: true,
-    extends: [],
-    env: {
-        browser: true,
-        es6: true,
-        node: true,
+  extends: [],
+  env: {
+    browser: true,
+    es6: true,
+    node: true,
+  },
+  parserOptions: {
+    createDefaultProgram: true,
+    project: 'tsconfig*.json',
+    sourceType: 'module',
+    errorOnUnknownASTType: true,
+    errorOnTypeScriptSyntacticAndSemanticIssues: true,
+    warnOnUnsupportedTypeScriptVersion: false,
+  },
+  overrides: [
+    {
+      files: ['*.ts'],
+      parser: '@typescript-eslint/parser',
+      plugins: ['sort-class-members', '@typescript-eslint'],
+      rules: {
+        ...require('./rules/base'),
+        ...require('./rules/import'),
+        ...require('./rules/member-ordering'),
+      },
     },
-    parserOptions: {
-        createDefaultProgram: true,
-        project: 'tsconfig*.json',
-        sourceType: 'module',
-        errorOnUnknownASTType: true,
-        errorOnTypeScriptSyntacticAndSemanticIssues: true,
-        warnOnUnsupportedTypeScriptVersion: false,
-    },
-    overrides: [
-        {
-            files: ['*.ts'],
-            parser: '@typescript-eslint/parser',
-            plugins: ['@typescript-eslint'],
-            rules: Object.assign(
-                require('./rules/eslint-plugin'),
-                require('./rules/import-plugin'),
-                require('./rules/typescript-eslint-plugin'),
-            ),
-        },
-    ],
+  ],
 };
