+++ packages/eslint-config-angular/package.json
@@ -18,11 +18,9 @@
     "type": "git",
     "url": "https://github.com/TinkoffCreditSystems/linters.git"
   },
-  "devDependencies": {
-    "@tinkoff/eslint-config": "^1.13.2"
-  },
   "peerDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@tinkoff/eslint-config": "^1.13.2",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
     "eslint": "^7.5.0",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-sort-class-members": "^1.7.0"
