+++ packages/eslint-config-angular/rules/base.js
@@ -1,0 +1,14 @@
+/**
+ * @note: only rules from eslint base rules
+ */
+module.exports = {
+  'no-param-reassign': 'error',
+  '@typescript-eslint/no-extraneous-class': ['error'],
+
+  /**
+   * @note: note you must disable the base rule
+   * as it can report incorrect errors in @typescript-eslint
+   */
+  'no-useless-constructor': 'off',
+  '@typescript-eslint/no-useless-constructor': ['error'],
+};
