--- packages/eslint-config-angular/rules/eslint-plugin.js
@@ -1,6 +1,0 @@
-/**
- * @note: only rules from eslint base rules
- */
-module.exports = {
-    'no-param-reassign': 'off', // todo: why off?
-};
