--- packages/eslint-config-angular/rules/import-plugin.js
@@ -1,6 +1,0 @@
-/**
- * @note: only rules which start with ^import/
- */
-module.exports = {
-    'import/no-webpack-loader-syntax': 'off',
-};
