+++ packages/eslint-config-angular/rules/import.js
@@ -1,0 +1,6 @@
+/**
+ * @note: only rules for imports
+ */
+module.exports = {
+  'import/no-webpack-loader-syntax': 'off',
+};
