--- packages/eslint-config-angular/rules/sort-class-members-plugin.js
@@ -1,9 +1,0 @@
-/**
- * @note: only rules which start with ^sort-class-members/
- */
-module.exports = {
-    /**
-     * @note(splincode): use @typescript-eslint/member-ordering instead
-     */
-    'sort-class-members/sort-class-members': 'off',
-};
