+++ packages/eslint-config-angular/test/member-ordering/__fixtures__/shared-utils.fixture.ts
@@ -1,0 +1,7 @@
+export function __Foo__() {
+    return (_target: unknown, _key: string) => {};
+}
+
+export function __Bar__() {
+    return (_target: unknown, _key: string) => {};
+}
