--- packages/eslint-plugin/tests/require-return-type.fixture.ts
@@ -1,119 +1,0 @@
-const Decorator = () => (_target: unknown, _key: string) => {};
-
-class A {
-    b() {}
-    public e(): string {
-        return '123';
-    }
-    protected e2(): string {
-        return '123';
-    }
-    private e3(): string {
-        return '123';
-    }
-    private e4(): string {
-        return '123';
-    }
-    static e5(): string {
-        return '123';
-    }
-    @Decorator()
-    d(): string {
-        return '123';
-    }
-    d2(a: string): string {
-        return '123';
-    }
-
-    c() {
-        // Should fail
-        return '123';
-    }
-    c2(/* 123 */) {
-        // Should fail
-        return '123';
-    }
-    c3() /* 123 */ {
-        // Should fail
-        return '123';
-    }
-    c4() {
-        // 321 // Should fail
-        return '123';
-    }
-    c5() {
-        return;
-    }
-}
-
-(() => {
-    new A();
-})();
-
-((): string => {
-    return '123';
-})();
-
-(() => {
-    return '123';
-})();
-
-const arrow = (): string => '123';
-const arrow2 = (param: string): string => param;
-const arrow3 = () => '123';
-const arrow4 = () => '123';
-const arrow5 = (param: string) =>
-    param +
-    function () {
-        new A();
-    };
-
-+(function (): string {
-    return '123';
-})();
-
-+(function () {
-    return '123';
-})();
-
-function ab() {
-    new A();
-}
-
-function ab2() {
-    return;
-}
-
-function ab3() {
-    // Should fail
-    return '123';
-}
-
-// Edge case
-// () => ({ action: 'xxx' }) as const
-
-class B {
-    constructor() {
-        ['a'].map(a => {
-            return 0;
-        });
-
-        ['a'].map(function (a): number {
-            return 0;
-        });
-
-        ['a'].map(function (a) {
-            if (a) {
-                return 0;
-            }
-
-            return 1;
-        });
-    }
-}
-
-function fnWithNestedFn() {
-    function nestedFn() {
-        return '';
-    }
-}
