+++ packages/linters/package.json
@@ -17,10 +17,10 @@
   },
   "dependencies": {
     "@tinkoff/eslint-plugin": "^1.13.2",
-    "@typescript-eslint/eslint-plugin": "^5.3.0",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.3.0",
-    "@typescript-eslint/experimental-utils": "^5.3.0",
-    "@typescript-eslint/parser": "^5.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.4.0",
+    "@typescript-eslint/experimental-utils": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "eslint": "^7.5.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-functional": "^3.7.2",
